requirements.txt
eth-brownie==1.18.1
numpy
pandas
scipy
tabulate
